requirements.txt
# Function dependencies, for example:
# package>=version
fastapi==0.68.1
aiofiles
uvicorn==0.15.0 
gunicorn==20.0.4
uvloop==0.15.2
python-jose==3.2.0
passlib==1.7.4
pydantic==1.8.2
python-multipart==0.0.5
bcrypt==3.2.0
numpy==1.19.2
#.....................#    
pandas==1.3.0
gspread==3.7.0
df2gspread==1.0.4
line-bot-sdk
DateTime
google-auth
google-cloud-storage
pandas-gbq==0.14.0